acromag: merge 235 and 236 packages
previous layout allowed either ap235 or ap236 to be included in a program,
not both.  This was due to duplicate copies of apcommon.c (not .h) included
in the final program, and a limitation of the cgo wrapper of the C linker.

This change allows them to co-exist, and also reduces the total amount
of code by deleting all of the duplicate code between them.

The vendor C files have been modified in compliance with the license to
add an include guard to apcommon.h, and change the includes for all of
the 235 and 236 C files to the localized path.